backend/djangoautotel_backend/requirements.txt
Django==4.2.3
djangorestframework
psycopg2-binary
